systemd: Add systemd service for EcoBot
This is required for running the bot on a server. Whenever
the server starts, this service is executed. It is under
the assumption that you have a user called "ecobot" and
that you clone the ecobot repository in the root of the ecobot user

Otherwise, the service will fire an error and the bot won't start.

Signed-off-by: kingbri <bdashore3@gmail.com>