botwide: Implement a simpler CRUD system
The system I was using was extremely complicated in terms of
understanding how it works, fetching balances, accounts, etc.

We now use the JSON.stringify method because this data is exteremely
simple and requires only one word/number.

The custom stringify function allows for arrays to be added, but
will be used for files that require storage in the form of an array.

Additionally, Fix inconsistency between using message.author.id and
message as arguments for functions. Every function takes the
message and message.author.id is used when we need a user id.

Signed-off-by: kingbri <bdashore3@gmail.com>